+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # develop
+  * Feature: NLSML building
   * Bugfix: Don't load rubygems because it is evil
 
 # 1.0.0 - 2012-03-13